# Reduce formatter's spec requires