@bcardiff 2017-03-27
Maybe we should split https://github.com/crystal-lang/crystal/blob/master/spec/spec_helper.cr#L1-L3 in another file and require that at the beginning of the `formatter_spec.cr`. That will also allow us to avoid the relative require of the formatter code. 

WDYT?