@MakeNowJust 2017-04-11
@bcardiff Totally good. I think we don't need discussion about it. But this pull request is self-contained. I think your suggestion can be separated from this pull request.