@miketheman 2017-04-23
#codetriage 

We last left this with the open question from @bcardiff about making another change to the `spec_helper.cr` file and @MakeNowJust responded with the detail that this PR seems to be self-contained and solves this.

I think that "relative requires" are part of the current game, until such a time where Crystal does auto-require-path-search, so I don't think there's anything inherently wrong with it.