@mverzilli 2017-04-23
Thanks @miketheman, I agree with your asessment. Let's merge this, the compile time gains are huge and the change is pretty hygienic. 7.5x better compile times just by being more specific declaring dependencies? Where do I sign? :).

@MakeNowJust this is an amazing contribution, thanks so much ❤️.